pyproject.toml
[tool.poetry]
name = "data-hub"
version = "0.1.0"
description = ""
authors = ["Eric Erdmann"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.20.0"
celery = "^5.2.7"
mongoengine = "^0.24.2"
django-viewflow = "^1.11.0"
django-material = "^1.11.3"
poetry-dotenv-plugin = "^0.1.0-alpha.2"
django-material-admin = "^1.8.6"
Django = "^4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
